feat: Refactor CAN speed processing and add AGL build script
This commit includes several updates to the application and build process.

CAN Data Processing:
- Refactors the CAN data handling for vehicle speed to process a 32-bit float instead of a 16-bit integer, aligning with an updated data format from the CAN bus.
- Updates VehicleData and CANReader to use  for speed values.
- Simplifies the CAN device initialization in , assuming the interface is pre-configured.

Build System:
- Adds a new cross-compilation script  for building the application for Automotive Grade Linux targets.
- The script sources the AGL SDK, uses the correct toolchain, and packages dependencies from the SDK's sysroot.

+++ qt_app/include/vehicledata.hpp
@@ -26,7 +26,7 @@ public:
     ~VehicleData() override;
 
     // Getters
-	double  getSpeed() const;
+	float   getSpeed() const;
 	double  getEnergy() const;
 	int     getBattery() const;
 	int     getDistance() const;
@@ -35,7 +35,7 @@ public:
     bool    getAutonomousMode() const;
 
     // Setters
-    void    setSpeed(double mps);
+    void    setSpeed(float mps);
     void    setEnergy(double energy);
     void    setBattery(int battery);
     void    setDistance(int distance);
@@ -75,7 +75,7 @@ private:
     void    markPropertyStale(const QString &propName);
 
     // Member variables
-    double  m_speed;
+    float   m_speed;
     double  m_energy;
     int     m_battery;
     int     m_distance;
